+++ src/deluge/processing/sound/sound_drum.h
@@ -64,4 +64,6 @@ public:
 	ArpeggiatorBase* getArp();
 	ArpeggiatorSettings* getArpSettings(InstrumentClip* clip = NULL) { return &arpSettings; }
 	void resetTimeEnteredState();
+
+	void stopDelay() override;
 };