+++ src/deluge/processing/sound/sound_instrument.cpp
@@ -87,6 +87,10 @@ void SoundInstrument::cutAllSound() {
 	Sound::unassignAllVoices();
 }
 
+void SoundInstrument::stopDelay() {
+	Sound::stopDelay();
+}
+
 void SoundInstrument::renderOutput(ModelStack* modelStack, StereoSample* startPos, StereoSample* endPos,
                                    int32_t numSamples, int32_t* reverbBuffer, int32_t reverbAmountAdjust,
                                    int32_t sideChainHitPending, bool shouldLimitDelayFeedback, bool isClipActive) {